Move the command to/from proto conversion out of the api package.
This simplifies the dependencies of the API package and avoids circles.
